Implement buffers.
Now when a file is opened, a buffer containing it is created on the fly.
As each file has their own buffer, it is not needed to read from disk
every time user wants to open a file. Also, unsaved changes won't be
lost when user changes buffer (TODO: implement function "Save All").

This commit also set up the bases to multiple kind of editors (= a file
can be open in a different editor depending of their mime-type).