Make /dev folder public
It happens that Service Worker do not trigger the Basic Auth
credential window when content fetched returns 401. Thus whenever
the browser "forgets" the credentials, playJS becomes unusable until
user unregister the service worker (or delete all the site's data on
mobile, which sucks as we want to have a working offline mode).
In fact playJS do not need to be private, as it is an open source
software. Thus the easiest workaround to this web browser (HTML5 spec ?)
is to make "/dev" folder public, but create a "/priv" directory where
user can store their private information.